+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 A major update to the BOOMER algorithm that introduces the following changes:
 
-* The parameter `--recalculate-predictions` allows to specify whether the predictions of rules should be recalculated on the entire training data, if instance sub-sampling is used.
+* The parameter `--instance-sub-sampling` does now allow to use stratified sampling (`stratified-label-wise` and `stratified-example-wise`).
+* The parameter `--recalculate-predictions` does now allow to specify whether the predictions of rules should be recalculated on the entire training data, if instance sub-sampling is used.
 
 ### Version 0.5.0 (Mar 31, 2021)